rename hnswlib_c, fix valgrind test
