+++ src/VecSim/vec_sim.cpp
@@ -2,7 +2,7 @@
 #include "VecSim/query_results.h"
 #include "VecSim/query_result_struct.h"
 #include "VecSim/algorithms/brute_force/brute_force.h"
-#include "VecSim/algorithms/hnsw/hnswlib_c.h"
+#include "VecSim/algorithms/hnsw/hnsw_wrapper.h"
 #include "VecSim/utils/vec_utils.h"
 #include "VecSim/utils/arr_cpp.h"
 #include <cassert>
